Progress on memory saftey, new newick parser
In order to remove the parent reference from the Node class, I need to
rewrite the newick parser. In addition, I added tests for the trees, and
everything seems to be passing.

A lot (but not all) of bare pointers have been changed to shared_ptr, in
the interest of memory saftey. It isn't clear what the performance
effect of this is, but I am more interested in being able to safely make
large refactors later safely.