Removed the parent field from the node class
This has been a long time coming. This field needs to go, because it
induces a weak circular depedancy to the internal nodes of a tree. In
order to solve this problem, the choice was made to remove the member
entirely. Using a weak_ptr is difficult because of the addChild
function. This is a member function that will add a child, and set the
parent in the child node. In order to make a weak_ptr, you need the
original shared_ptr. Since a node doesn't know of its own shared_ptr,
this function would have to be an independant function. The choice was
mad that it was easier to change how some algorithms are performed,
instead of going through the code and changing all instances of
addChild.

This might not be the best way to do this, but many of the algorithms
that have been changed really were not good, so the work there was not
wasted. For a few operations, it might be better to re-add _parent as a
weak_ptr, so that we have faster access to the parent member. In order
to find out, we will need to perform benchmarks, but I suspect that this
will be a very minor loss.