Distributions are now 64bit ints instead of vectors.
The biggest implication is that going forward, we need to be careful of
the order of the ranges. It should be pretty easy, as dist & (1 << i) is
the ith range. This does mean that when we print out the number in
decimal, the endianness is flipped. Be aware of this