Refactored the config file code and the alignment parser
This commit has three changes features included:
1. A reimplementation of the config file parser
2. A reimplementation of alignment parser
3. Removal of InputReader as a class and a file.

The config file parsing is now done with a grammar, which allows for
better error messages. This means that the error messages should be
more descriptive, at least in terms of the location of the error.

The alignment parser is now mostly based on std::istream builtin
functions, and is much simpler. Additionally, fasta files are now
supported.

Finally, both of these changes made it convenient to remove the
InputReader class, which was always a pseudo namespace in reality.