Implement fixed nodes for fossil support
In order to support fossile nodes, there is now an option to fix a node
to a specific distribution only. When a node has a fixed distribution,
conditional likelihoods other than the fixed distribution are set to
zero. In the case of the root, the prior frequencies are also set to
zero except for the fixed node (though, this is superfluous).

Additionally, reverse operations also only compute the fixed dist value.

Finally, there is a bug fix in the case of limited max areas.