Change the behavior of the fossil option to match the original behavior
Lagrange (the old version) implemented fossils as a constrant on the
node. I instead implemented them as a fixed node, which was another
option. This commit restores that old behavior, and includes the new
behavior under a `fixed` sub option for fossil.

Additionally, some initial support was added for the implementation of
different time periods.