fix(periods): fix an issue with zero length branches
In the case that there are zero length branches, the period length is
also zero. This causes some internal issues with operation generation.
The solution here is to add a very small length to a zero length period,
but this is a bit of a hack. In the future, a better solution should be
found.