feat(spilts): Compute splits for specific nodes only
Given the size of results for splits, it is useful to only compute a
select few splits. Now, the ability to do this has been implemented

As part of this however, a lot of technical debt needed to be paid off.
Now, results are stored in the tree datastructure, and the results are
pulled directly from the tree. This is much simpler in terms of code,
and should allow me to examine results more clearly in the future.