feat(fossil): Support `include` and `exclude` fossils
Adds support for 2 new types of fossils: `include` and `exclude`. Both
fossils constrain a particular node to a set of states, based on a mask.

The code to do this co-opts the fossil mechanism to do this, and so the
amount of modification is limited. Additionally, most of the code to do
this in `Operation.cpp` already existed. All this commit really does is
"plug it in".