feat(opt): Implement new optimization methods
NLopt has a ton of optimization methods that just work. The choice of
Nelder-Mead in the original release of `lagrange-ng` was just to match
parity with the original `lagrange-cpp`. However, it seems that choice
was arbitrary, as other optimization methods appear to work better. In
particular, BFGS works well when the number of parameters (periods) is
high.

This commit breaks with `lagrange-cpp` and sets the new default to be
BFGS. In small cases (one period), it is a bit slower as it has to
compute the gradient. But in cases with more periods, it is
significantly faster.