fix(splits): Fix an error when computing only splits
When we compute the splits tree, we condition the split at the node
based on the best ancestral state. This means that for every split we
compute, we also need to compute the state. This commit introduces
logic to ensure this happens, but keeps the output restricted to what
the user asked for.

Additionally, this commit changes some internal things. State and split
goals now keep the node ID as part of the structure. This ID is used to
match the results up with the node in the tree data structure, so that
we can print the results out correctly. To this end, results are now
stored in an `unordered_map` from `size_t` to whatever result is used.
We traverse the tree, and assign results if the node ID is in the map.

State and split nodes indicated by MRCAs in the config file are now
stored as a set, so that we can merge them.

All together, this simplifies the code path that assigns results to the
tree, as we now don't have a difference between some states/splits and
all state/splits cases. Additionally, we no longer have a strong
requirement on traversal order and the result vector order. Finally, the
set data structure takes care of the user doing weird things like
specifying the same node multiple times.