feat(ambig): Add code to coerce invalid tip data
In the case that the user provides a tip range that is too large for the
max areas set, we now coerce that tip state into an ambiguous state that
is the closest that it can be, in a hypercube sense.

Specifically, it now sets every state that is
- A strict subset of the range, and
- Equal to the number of max areas.

For example, 111 with max areas 2 becomes (011, 101, 110).

In addition, we warn the user that the data is is being coerced