fix: Fix bugs relating to the reverse operations
There was a couple of issues with the reverse operations. The first was
a race condition where a dispersion operation could be executed twice.
The second was that the generated reverse graph would skip over period
boundries.

Both of these have been fixed. The race condition was fixed by making
disperison operations check if they actually need to be executed before
running the rest of the code. This is not an ideal solution, but it is
fine for now.

The second bug was fixed by simply making the graph traversal correct.
However, in order to determine that there was an error, I implemented a
debug mode which will dump the forward and reverse graphs to graphviz,
which can then be made into an image. The information is pretty
primative, but it still is super helpful. The option to enable this in
the config file is `dump-graph`.