fix(sync): resolve the issue with threads locking when halting
There was an issue with threads deadlocking when the halt was given. The
issue was caused by the barrier being too slow to release threads, which
caused them to execute the switch body with an "early" mode, before
other threads could get the new mode. The early thread would then halt,
and never encounter the barrier.

The solution to this is to add a second barrier to the end of the main
work loop body, ensuring that everybody does the loop exactly once
before reassessing the next loop iteration. So, there is a barrier as
every thread waits for the others to finish, and then a barrier while
all the non-main threads wait for the main thread to decide what to do.