feat: Implement CI using github actions
As a part of this, I changed how the BLAS dependencies are resolved.
This is so that we can accelerate the build for testing, and for static
builds we just build openblas from source.

In addition, I had to fix the issues with building on apple anyways,
which ended up defeating the point of this whole exercise. The solution
in this case is to use preprocessor if statements and feature test
macros to conditionally include code which does not use the new
features. The OSX code passes all tests, but I haven't rigorously
tested it.