feat: Implement checkpointing
Checkpointing is simply done by saving the parameters to a CSV file as
the program is running. The current rate is about every 10 iterations,
or every 10 seconds.

When the we detect that there is already a checkpoint, we will load the
file and either finish optimization or just evaluate the tree so that we
can compute splits/states.